A09: Correciones backend

+++ src/controllers/ProjectController.ts
@@ -2,9 +2,13 @@ import type { Request, Response } from 'express';
 import Project from '../models/Project';
 
 export class ProjectController {
-// Crear un nuevo proyecto
+    // Crear un nuevo proyecto
     static async createProject(req: Request, res: Response) {
         const project = new Project(req.body);
+
+        // Asigna manager
+        project.manager = req.user.id
+
         try {
             await project.save();
             res.send("Proyecto creado correctamente");
@@ -13,23 +17,33 @@ export class ProjectController {
             res.status(500).send("Error al crear el proyecto");
         }
     }
-// Obtener todos los proyects
+    // Obtener todos los proyects
     static async getAllProjects(req: Request, res: Response) {
         try {
-            const projects = await Project.find();
+            const projects = await Project.find({
+                $or: [
+                    { manager: { $in: req.user.id } }
+                ]
+            });
             res.json(projects)
         } catch (error) {
             console.log(error);
             res.status(500).send("Error al obtener los proyectos");
         }
     }
-// Obtener un proyecto por ID
+    // Obtener un proyecto por ID
     static async getProjectById(req: Request, res: Response) {
         const { id } = req.params;
         try {
             const project = await Project.findById(id).populate('tasks');
             if (!project) {
-                return res.status(404).send("Proyecto no encontrado");
+                const error = new Error('Proyecto no encontrado')
+                return res.status(404).json({ error: error.message })
+            }
+
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción no valida')
+                return res.status(404).json({ error: error.message })
             }
             res.json(project);
         } catch (error) {
@@ -37,7 +51,7 @@ export class ProjectController {
             res.status(500).send("Error al obtener el proyecto");
         }
     }
-// Actualizar un proyecto por ID
+    // Actualizar un proyecto por ID
     static async updateProjectById(req: Request, res: Response) {
         const { id } = req.params;
         try {
@@ -45,6 +59,12 @@ export class ProjectController {
             if (!project) {
                 return res.status(404).send("Proyecto no encontrado");
             }
+
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Solo el Manager puede actualizar un proyecto')
+                return res.status(404).json({ error: error.message })
+            }
+
             project.projectName = req.body.projectName;
             project.clientName = req.body.clientName;
             project.description = req.body.description;
@@ -56,7 +76,7 @@ export class ProjectController {
         }
     }
 
-// Eliminar un proyecto por ID
+    // Eliminar un proyecto por ID
     static deleteProjectById = async (req: Request, res: Response) => {
         const { id } = req.params;
         try {
@@ -64,9 +84,14 @@ export class ProjectController {
 
             if (!project) {
                 const error = new Error("Proyecto no encontrado");
-                return res.status(404).json({error: error.message});
+                return res.status(404).json({ error: error.message });
             }
-            
+
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Solo el Manager puede eliminar el proyecto')
+                return res.status(404).json({ error: error.message })
+            }
+
             await project.deleteOne();
             res.send("Proyecto eliminado correctamente");
         } catch (error) {
